Add dataset interface and attention dataset implementation.
