+++ libs/preprocessing/tests/test_attention_dataset.py
@@ -2,11 +2,11 @@
 # Unit tests for extractor
 #
 
-import numpy as np
 import unittest
 
-from attention_for_histone_modification.libs.preprocessing.attention_types import (
-        AttentionDataset, AttentionDatasetConfig, AttentionTrainingExample)
+from attention_for_histone_modification.libs.preprocessing.attention_training_example import AttentionTrainingExample
+from attention_for_histone_modification.libs.preprocessing.attention_dataset import (
+        AttentionDataset, AttentionDatasetConfig)
 from attention_for_histone_modification.libs.preprocessing.tests.utilities_for_tests import (
         create_attention_config_by_indices, create_training_example_by_label)
 
@@ -33,7 +33,6 @@ class TestAttentionDataset(unittest.TestCase):
         indexed_training_examples = self.dataset.get_training_examples(query_indices)
         for (index, te) in indexed_training_examples:
             self.assertEqual(index, te.label)
-
 
 if __name__ == '__main__':
     unittest.main()
